refactor/Move function createOrUpdateAtm to te Atm provider
