+++ obp-api/src/main/scala/code/atms/MappedAtmsProvider.scala
@@ -1,9 +1,13 @@
 package code.atms
 
 import code.api.util.{OBPLimit, OBPOffset, OBPQueryParam}
+import code.bankconnectors.LocalMappedConnector.getAtmLegacy
+import code.util.Helper.optionBooleanToString
 import code.util.{TwentyFourHourClockString, UUIDString}
 import com.openbankproject.commons.model._
+import net.liftweb.common.{Box, Full}
 import net.liftweb.mapper._
+import net.liftweb.util.Helpers.tryo
 
 import scala.collection.immutable.List
 
@@ -23,6 +27,132 @@ object MappedAtmsProvider extends AtmsProvider {
     Some(MappedAtm.findAll(mapperParams:_*))
   }
 
+  override def createOrUpdateAtm(atm: AtmT): Box[AtmT] = {
+
+    val isAccessibleString = optionBooleanToString(atm.isAccessible)
+    val hasDepositCapabilityString = optionBooleanToString(atm.hasDepositCapability)
+    val supportedLanguagesString = atm.supportedLanguages.map(_.mkString(",")).getOrElse("")
+    val servicesString = atm.services.map(_.mkString(",")).getOrElse("")
+    val accessibilityFeaturesString = atm.accessibilityFeatures.map(_.mkString(",")).getOrElse("")
+    val supportedCurrenciesString = atm.supportedCurrencies.map(_.mkString(",")).getOrElse("")
+    val notesString = atm.notes.map(_.mkString(",")).getOrElse("")
+    val locationCategoriesString = atm.locationCategories.map(_.mkString(",")).getOrElse("")
+
+    //check the atm existence and update or insert data
+    getAtmLegacy(atm.bankId, atm.atmId) match {
+      case Full(mappedAtm: MappedAtm) =>
+        tryo {
+          mappedAtm.mName(atm.name)
+            .mLine1(atm.address.line1)
+            .mLine2(atm.address.line2)
+            .mLine3(atm.address.line3)
+            .mCity(atm.address.city)
+            .mCounty(atm.address.county.getOrElse(""))
+            .mCountryCode(atm.address.countryCode)
+            .mState(atm.address.state)
+            .mPostCode(atm.address.postCode)
+            .mlocationLatitude(atm.location.latitude)
+            .mlocationLongitude(atm.location.longitude)
+            .mLicenseId(atm.meta.license.id)
+            .mLicenseName(atm.meta.license.name)
+            .mOpeningTimeOnMonday(atm.OpeningTimeOnMonday.orNull)
+            .mClosingTimeOnMonday(atm.ClosingTimeOnMonday.orNull)
+
+            .mOpeningTimeOnTuesday(atm.OpeningTimeOnTuesday.orNull)
+            .mClosingTimeOnTuesday(atm.ClosingTimeOnTuesday.orNull)
+
+            .mOpeningTimeOnWednesday(atm.OpeningTimeOnWednesday.orNull)
+            .mClosingTimeOnWednesday(atm.ClosingTimeOnWednesday.orNull)
+
+            .mOpeningTimeOnThursday(atm.OpeningTimeOnThursday.orNull)
+            .mClosingTimeOnThursday(atm.ClosingTimeOnThursday.orNull)
+
+            .mOpeningTimeOnFriday(atm.OpeningTimeOnFriday.orNull)
+            .mClosingTimeOnFriday(atm.ClosingTimeOnFriday.orNull)
+
+            .mOpeningTimeOnSaturday(atm.OpeningTimeOnSaturday.orNull)
+            .mClosingTimeOnSaturday(atm.ClosingTimeOnSaturday.orNull)
+
+            .mOpeningTimeOnSunday(atm.OpeningTimeOnSunday.orNull)
+            .mClosingTimeOnSunday(atm.ClosingTimeOnSunday.orNull)
+            .mIsAccessible(isAccessibleString) // Easy access for people who use wheelchairs etc. Tristate boolean "Y"=true "N"=false ""=Unknown
+            .mLocatedAt(atm.locatedAt.orNull)
+            .mMoreInfo(atm.moreInfo.orNull)
+            .mHasDepositCapability(hasDepositCapabilityString)
+            .mSupportedLanguages(supportedLanguagesString)
+            .mServices(servicesString)
+            .mNotes(notesString)
+            .mAccessibilityFeatures(accessibilityFeaturesString)
+            .mSupportedCurrencies(supportedCurrenciesString)
+            .mLocationCategories(locationCategoriesString)
+            .mMinimumWithdrawal(atm.minimumWithdrawal.orNull)
+            .mBranchIdentification(atm.branchIdentification.orNull)
+            .mSiteIdentification(atm.siteIdentification.orNull)
+            .mSiteName(atm.siteName.orNull)
+            .mCashWithdrawalNationalFee(atm.cashWithdrawalNationalFee.orNull)
+            .mCashWithdrawalInternationalFee(atm.cashWithdrawalInternationalFee.orNull)
+            .mBalanceInquiryFee(atm.balanceInquiryFee.orNull)
+            .saveMe()
+        }
+      case _ =>
+        tryo {
+          MappedAtm.create
+            .mAtmId(atm.atmId.value)
+            .mBankId(atm.bankId.value)
+            .mName(atm.name)
+            .mLine1(atm.address.line1)
+            .mLine2(atm.address.line2)
+            .mLine3(atm.address.line3)
+            .mCity(atm.address.city)
+            .mCounty(atm.address.county.getOrElse(""))
+            .mCountryCode(atm.address.countryCode)
+            .mState(atm.address.state)
+            .mPostCode(atm.address.postCode)
+            .mlocationLatitude(atm.location.latitude)
+            .mlocationLongitude(atm.location.longitude)
+            .mLicenseId(atm.meta.license.id)
+            .mLicenseName(atm.meta.license.name)
+            .mOpeningTimeOnMonday(atm.OpeningTimeOnMonday.orNull)
+            .mClosingTimeOnMonday(atm.ClosingTimeOnMonday.orNull)
+
+            .mOpeningTimeOnTuesday(atm.OpeningTimeOnTuesday.orNull)
+            .mClosingTimeOnTuesday(atm.ClosingTimeOnTuesday.orNull)
+
+            .mOpeningTimeOnWednesday(atm.OpeningTimeOnWednesday.orNull)
+            .mClosingTimeOnWednesday(atm.ClosingTimeOnWednesday.orNull)
+
+            .mOpeningTimeOnThursday(atm.OpeningTimeOnThursday.orNull)
+            .mClosingTimeOnThursday(atm.ClosingTimeOnThursday.orNull)
+
+            .mOpeningTimeOnFriday(atm.OpeningTimeOnFriday.orNull)
+            .mClosingTimeOnFriday(atm.ClosingTimeOnFriday.orNull)
+
+            .mOpeningTimeOnSaturday(atm.OpeningTimeOnSaturday.orNull)
+            .mClosingTimeOnSaturday(atm.ClosingTimeOnSaturday.orNull)
+
+            .mOpeningTimeOnSunday(atm.OpeningTimeOnSunday.orNull)
+            .mClosingTimeOnSunday(atm.ClosingTimeOnSunday.orNull)
+            .mIsAccessible(isAccessibleString) // Easy access for people who use wheelchairs etc. Tristate boolean "Y"=true "N"=false ""=Unknown
+            .mLocatedAt(atm.locatedAt.orNull)
+            .mMoreInfo(atm.moreInfo.orNull)
+            .mHasDepositCapability(hasDepositCapabilityString)
+            .mSupportedLanguages(supportedLanguagesString)
+            .mServices(servicesString)
+            .mNotes(notesString)
+            .mAccessibilityFeatures(accessibilityFeaturesString)
+            .mSupportedCurrencies(supportedCurrenciesString)
+            .mLocationCategories(locationCategoriesString)
+            .mMinimumWithdrawal(atm.minimumWithdrawal.orNull)
+            .mBranchIdentification(atm.branchIdentification.orNull)
+            .mSiteIdentification(atm.siteIdentification.orNull)
+            .mSiteName(atm.siteName.orNull)
+            .mCashWithdrawalNationalFee(atm.cashWithdrawalNationalFee.orNull)
+            .mCashWithdrawalInternationalFee(atm.cashWithdrawalInternationalFee.orNull)
+            .mBalanceInquiryFee(atm.balanceInquiryFee.orNull)
+            .saveMe()
+        }
+    }
+  }
 
 }
